PostgreSQL support (thx abraxas)
In order to enable postgresql support, ./install must be called and postgresql given
as the db backend. The oracle folder will then be removed and the BaseEventDbEvent of
pgsql used. Here the comments 'user' column is mapped to 'username' and IS_POSTGRESQL
is a property that will be checked to bypass address resolution (not nice, but currently
does the job)

refs #1683