Fix variable scope of `selectedRecords'
The comment functionality of the event's detail panel, cannot use the
`eventGrid' as source for the selected records since checking rows
via the row's checkbox is different to selecting a single row.
The `eventGrid` is a parameter to the function `CommentForm::show()'
which creates the comment window only once and references the `eventGrid'
parameter in its submit handler. Because the window is only created once
the selected records cannot be reset via a function parameter.

The most simple approach would be to recreate the comment window
every time but the selected records are now a property of the
comment window which is set on construction the first time `CommentForm::show()'
is called and overriden when it is already created.